Update libft : ft_memset

+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: yonyoo <yonyoo@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/17 12:03:40 by yonyoo            #+#    #+#             */
-/*   Updated: 2023/03/17 12:06:04 by yonyoo           ###   ########.fr       */
+/*   Updated: 2023/03/17 12:08:22 by yonyoo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 void	*memset(void *b, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
